refactor: reuse CodeMirror for CSS editors
