+++ apps/docs/docs/advanced/styling/index.mdx
@@ -85,8 +85,11 @@ To do this, enter the board settings:
 
 ![](./img/enter-board-settings.png)
 
-Scroll to the **Custom CSS** section on the board settings page. It is part of the unified settings form; enter your CSS
-there and save the board settings to apply it to that board.
+Administrators can add instance-wide CSS under **Management → Settings → Branding → Custom CSS**. Enter CSS directly
+or import a community theme from the Workshop, then save the server settings to apply it to every page.
+
+For board-specific styling, scroll to the **Custom CSS** section on the board settings page. Enter your CSS there and
+save the board settings to apply it to that board. Board CSS is loaded after the global stylesheet and can override it.
 
 ![](./img/custom-css-setting.png)
