+++ apps/nextjs/src/components/custom-css/custom-css-editor.tsx
@@ -1,12 +1,6 @@
 "use client";
 
-import { Group, Input, Stack } from "@mantine/core";
-import { highlight, languages } from "prismjs";
-import Editor from "react-simple-code-editor";
-
-import "~/styles/prismjs.scss";
-
-import classes from "./custom-css-editor.module.css";
+import { CodeEditor } from "~/components/custom-widgets/code-editor";
 
 interface CustomCssEditorProps {
   id: string;
@@ -28,26 +22,17 @@ export const CustomCssEditor = ({
   actions,
 }: CustomCssEditorProps) => {
   return (
-    <Stack gap="xs">
-      <Group justify="space-between" align="end">
-        <div>
-          <Input.Label htmlFor={id}>{label}</Input.Label>
-          <Input.Description>{description}</Input.Description>
-        </div>
-        {actions}
-      </Group>
-      <Input.Wrapper error={error} inputWrapperOrder={["input", "error"]}>
-        <div className={classes.root}>
-          <Editor
-            textareaId={id}
-            onValueChange={onChange}
-            value={value}
-            highlight={(code) => highlight(code, languages.extend("css", {}), "css")}
-            padding={10}
-            className={classes.editor}
-          />
-        </div>
-      </Input.Wrapper>
-    </Stack>
+    <CodeEditor
+      id={id}
+      label={label}
+      description={description}
+      value={value}
+      onChange={onChange}
+      error={error}
+      language="css"
+      height="250px"
+      maxLength={16_384}
+      actions={actions}
+    />
   );
 };
